Header name must be same
Based on the logged data, it seems that the data array contains an object with a key attendances, which is an array of arrays of RowDataPacket objects. This structure does not match the keys in your desiredColumns array, which is why you're getting the "Missing property 'shift_name'" error.

To fix this issue, you need to access the attendances array and then map over each RowDataPacket object to extract the relevant data.